stat_svc - ver. 0.24 (ViewStats moved from model to dto)
