+++ statistics/statistics-dto/src/main/java/ru/practicum/explore_with_me/statistics_dto/dto/ViewStats.java
@@ -1,4 +1,4 @@
-package ru.practicum.explore_with_me.statistics_model;
+package ru.practicum.explore_with_me.statistics_dto.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
