+++ statistics/statistics_service/src/main/java/ru/practicum/explore_with_me/statistics_controller/StatisticsController.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.explore_with_me.statistics_dto.dto.EndpointHitDto;
-import ru.practicum.explore_with_me.statistics_model.ViewStats;
+import ru.practicum.explore_with_me.statistics_dto.dto.ViewStats;
 import ru.practicum.explore_with_me.statistics_service.StatisticsService;
 
 import java.util.List;
