+++ statistics/statistics_service/src/main/java/ru/practicum/explore_with_me/statistics_repository/StatisticsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ru.practicum.explore_with_me.statistics_model.EndpointHit;
-import ru.practicum.explore_with_me.statistics_model.ViewStats;
+import ru.practicum.explore_with_me.statistics_dto.dto.ViewStats;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -13,7 +13,7 @@ import java.util.List;
 @Repository
 public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
 
-    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_model.ViewStats(e.app, e.uri, COUNT(DISTINCT e.ip)) " +
+    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_dto.dto.ViewStats(e.app, e.uri, COUNT(DISTINCT e.ip)) " +
             "FROM EndpointHit e " +
             "WHERE " +
             "e.timestamp BETWEEN :start AND :end " +
@@ -24,7 +24,7 @@ public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
                                                    @Param("end") LocalDateTime end,
                                                    @Param("uris") List<String> uris);
 
-    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_model.ViewStats(e.app, e.uri, COUNT(e.ip)) " +
+    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_dto.dto.ViewStats(e.app, e.uri, COUNT(e.ip)) " +
             "FROM EndpointHit e " +
             "WHERE " +
             "e.timestamp BETWEEN :start AND :end " +
@@ -35,7 +35,7 @@ public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
                                                       @Param("end") LocalDateTime end,
                                                       @Param("uris") List<String> uris);
 
-    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_model.ViewStats(e.app, e.uri, COUNT(DISTINCT e.ip)) " +
+    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_dto.dto.ViewStats(e.app, e.uri, COUNT(DISTINCT e.ip)) " +
             "FROM EndpointHit e " +
             "WHERE " +
             "e.timestamp BETWEEN :start AND :end " +
@@ -44,7 +44,7 @@ public interface StatisticsRepository extends JpaRepository<EndpointHit, Long> {
     List<ViewStats> findAllUniqueWhenUriIsEmpty(@Param("start") LocalDateTime start,
                                                 @Param("end") LocalDateTime end);
 
-    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_model.ViewStats(e.app, e.uri,COUNT(DISTINCT e.ip)) " +
+    @Query("SELECT NEW ru.practicum.explore_with_me.statistics_dto.dto.ViewStats(e.app, e.uri,COUNT(DISTINCT e.ip)) " +
             "FROM EndpointHit e " +
             "WHERE " +
             "e.timestamp BETWEEN :start AND :end " +
